+++ util.h
@@ -24,7 +24,7 @@ inline uint_least32_t find_representative(
 }
 
 inline float manhattan(const pathfinder::Coord& a, const pathfinder::Coord& b) {
-  return pow(pow(std::abs((float)a.row - (float)b.row), 2) + pow(std::abs((float)a.col - (float)b.col), 2), 0.5);
+  return (float)pow(pow(std::abs((float)a.row - (float)b.row), 2) + pow(std::abs((float)a.col - (float)b.col), 2), 0.5);
 }
 
 }  // namespace util
